# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Plataforma de gestión de asistencia, turnos, sucursales, accesos y dispositivos",
        "title": "Reloj Control API",
        "termsOfService": "https://minolsoft.cl/terms",
        "contact": {
            "name": "Minolsoft",
            "url": "https://minolsoft.cl",
            "email": "soporte@minolsoft.cl"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://minolsoft.cl/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/shifts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET lista turnos. POST crea un nuevo turno.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shifts"
                ],
                "summary": "Turnos",
                "parameters": [
                    {
                        "description": "Crear turno",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.createShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ShiftDTO"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShiftDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET lista turnos. POST crea un nuevo turno.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shifts"
                ],
                "summary": "Turnos",
                "parameters": [
                    {
                        "description": "Crear turno",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.createShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ShiftDTO"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShiftDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/shifts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET obtiene un turno. PATCH actualiza parcialmente. DELETE elimina.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shifts"
                ],
                "summary": "Turno por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Actualizar turno",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.patchShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShiftDTO"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET obtiene un turno. PATCH actualiza parcialmente. DELETE elimina.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shifts"
                ],
                "summary": "Turno por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Actualizar turno",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.patchShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShiftDTO"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET obtiene un turno. PATCH actualiza parcialmente. DELETE elimina.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shifts"
                ],
                "summary": "Turno por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Actualizar turno",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.patchShiftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShiftDTO"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
...
        "handlers.ShiftDTO": {
            "type": "object",
            "properties": {
                "break_minutes": {
                    "type": "integer",
                    "example": 60
                },
                "created_at": {
                    "type": "string"
                },
                "crosses_midnight": {
                    "type": "boolean",
                    "example": false
                },
                "description": {
                    "type": "string",
                    "example": "Lunes a viernes 08:00 a 17:00"
                },
                "end_time": {
                    "type": "string",
                    "example": "17:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Turno mañana"
                },
                "start_time": {
                    "type": "string",
                    "example": "08:00"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "handlers.createShiftRequest": {
            "type": "object",
            "properties": {
                "break_minutes": {
                    "type": "integer",
                    "example": 60
                },
                "crosses_midnight": {
                    "type": "boolean",
                    "example": false
                },
                "description": {
                    "type": "string",
                    "example": "Lunes a viernes 08:00 a 17:00"
                },
                "end_time": {
                    "type": "string",
                    "example": "17:00"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Turno mañana"
                },
                "start_time": {
                    "type": "string",
                    "example": "08:00"
                }
            }
        },
...
        "handlers.patchShiftRequest": {
            "type": "object",
            "properties": {
                "break_minutes": {
                    "type": "integer",
                    "example": 60
                },
                "crosses_midnight": {
                    "type": "boolean",
                    "example": false
                },
                "description": {
                    "type": "string",
                    "example": "Lunes a viernes 08:00 a 17:00"
                },
                "end_time": {
                    "type": "string",
                    "example": "17:00"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Turno mañana"
                },
                "start_time": {
                    "type": "string",
                    "example": "08:00"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}